requirements.txt
transformers
torch
pytesseract
pillow
